# internal/api/api.swagger.yaml
swagger: '2.0'
info:
  title: Space VPX Backend Service
  version: development
  description: Space VPX WebSite Backend Service
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /solar_panel_side:
    post:
      security:
        - Bearer: []
      summary: Create Solar Panel Side
      operationId: CreateSolarPanelSide
      tags: 
        - SolarPanelSide
      parameters:
        - name: CreateSolarPanelSideBody
          description: Create Solar Panel Side Body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateSolarPanelSideBody'
      responses:
        200:
          description: Create Solar Panel Side Response
          schema:
            $ref: '#/definitions/SolarPanelSide'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/Error' 
        422:
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/Error"        
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
  /solar_panel_side/available_solar_panel_side:
    get:
      security:
        - Bearer: []
      summary: Get Cube Sat Solar Panels Side
      operationId: GetCubeSatSolarPanelsSide
      tags: 
        - SolarPanelSide
      parameters:
      - in: query
        required: true
        name: offset
        type: integer
        description: Offset Configs
        minimum: 0
        required: true
      - in: query
        required: true
        name: limit
        type: integer
        description: Offset Configs
        minimum: 0
        required: true
      - in: query
        name: sort[field]
        description: sort parametrs
        enum: [created_at]
        type: string
      - in: query
        name: FilterCubeSatSolarPanelByLength[min]
        description: Filter By Min Length
        type: number
      - in: query
        name: FilterCubeSatSolarPanelByLength[max]
        description: Filter By Max Length
        type: number
      responses:
        200:
          description: Get CubeSat Solar Panels Side Response
          schema:
            $ref: '#/definitions/CubeSatSolarPanelsSide'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/Error' 
        422:
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/Error"        
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'

### CUBE SAT FRAME
...
definitions:
...
  Error:
    type: object
    required:
      - message
    properties:
      message:
        type: string
...
  SolarPanelSide:
    type: object
    required:
      - id
    properties:
      id:
        type: integer
        format: int64
        description: Unique identifier for the chassis (auto-generated)
      name:
        type: string
      length:
        type: number
        description: Name of the chassis
      width:
        type: number
        description: Number of slots in the chassis
      height:
        type: number
        description: Size of the chassis
      max_operating_temperature:
        type: number
        description: Maximum operating temperature in degrees Celsius
      min_operating_temperature:
        type: number
        description: Minimum operating temperature in degrees Celsius
      weight:
        type: number
        description: Maximum non-operating temperature in degrees Celsius
      interface:
        type: string
        description: Minimum non-operating temperature in degrees Celsius
      voc:
        type: number
        description: Overload capacity
      isc:
        type: number
        description: Maximum sine vibration
      vmp:
        type: number
        description: Minimum sine vibration
      imp:
        type: number
        description: Maximum random vibration
      efficiency:
        type: integer
        description: Minimum random vibration
      coil_area:
        type: number
        description: Number of axes
      coil_resistance:
        type: integer
        description: Shock response spectrum value
      mechanical_vibration:
        type: integer
        description: Peak overload spectrum value 1
      mechanical_shock:
        type: integer
        description: Peak overload spectrum value 2
      updated_at:
        type: integer
        description: Timestamp of the last update
      created_at:
        type: integer
        description: Timestamp of creation
  CubeSatSolarPanelsSide:
    type: object
    required:
      - count
    properties:
      count:
        type: integer
      cube_sat_solar_panels_side:
        type: array
        items:
          $ref: "#/definitions/SolarPanelSide"
  CreateSolarPanelSideBody:
    type: object
    required:
      - length
      - width
      - height
    properties:
      name:
        type: string
      length:
        type: number
      width:
        type: number
      height:
        type: number
      max_operating_temperature:
        type: number
      min_operating_temperature:
        type: number
      weight:
        type: number
      interface:
        type: string
      voc:
        type: number
      isc:
        type: number
      vmp:
        type: number
      imp:
        type: number
      efficiency:
        type: integer
      coil_area:
        type: number
      coil_resistance:
        type: integer
      mechanical_vibration:
        type: integer
      mechanical_shock:
        type: integer
...